Move abstract constructs to its own subfolder
