+++ src/constructs/abstracts/index.ts
@@ -1,0 +1,1 @@
+export { AwsConstruct } from "./AwsConstruct";
