+++ src/constructs/aws/Webhook.ts
@@ -5,7 +5,7 @@ import { EventBus } from "@aws-cdk/aws-events";
 import { FromSchema } from "json-schema-to-ts";
 import { PolicyDocument, PolicyStatement, Role, ServicePrincipal } from "@aws-cdk/aws-iam";
 import { AwsProvider } from "@lift/providers";
-import { AwsConstruct } from "@lift/constructs/aws";
+import { AwsConstruct } from "@lift/constructs/abstracts";
 import ServerlessError from "../../utils/error";
 
 const WEBHOOK_DEFINITION = {
